requirements.txt
fastapi~=0.61.1
SQLAlchemy~=1.3.19
pydantic~=1.6.1
uvicorn~=0.11.8
python-jose[cryptography]~=3.2.0
passlib~=1.7.2
argon2_cffi==20.1.0
python-dotenv==0.14.0
yeelight~=0.5.3